Convert to new concordance method
