+++ lifelines/tests/test_estimation.py
@@ -618,7 +618,7 @@ prio  2.639e-01  1.302e+00 8.291e-02  3.182e+00 1.460e-03   1.013e-01   4.264e-0
 ---
 Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
 
-Concordance = 0.644""".strip().split()
+Concordance = 0.640""".strip().split()
             for i in [0, 1, 2, -2, -1]:
                 assert output[i] == expected[i]
         finally:
